Cleanup orphaned using ids, not timestamp sorting
